Instantiate ".sum" of int array template in phobos for testing
This is just a hack for testing. Calculating a sum of int array
may overflow and abort the application with -ftrapv option.

The application and the phobos library may be compiled in different
mismatched -ftrapv configurations, so it makes sense to be able
to test how templates instantiation is affected.